minor refactor

+++ Include/Game/Tilemap.hpp
@@ -35,7 +35,7 @@ class Tilemap : public SceneNode
 
 		void							getRoom(Tile::ID id, std::vector<TilePtr>& room);
 		void							getRoom(sf::Vector2f position, std::vector<TilePtr>& room);
-		sf::Vector2i 					getRandomRoomCenter();
+		sf::Vector2f 					getRandomRoomCenter();
 
 
 	private:
